+++ mite_web/mite_web/config/extensions.py
@@ -1,4 +1,5 @@
 from flask_restx import Api
+from flask_sqlalchemy import SQLAlchemy
 
 api = Api(
     title="MITE API",
@@ -6,3 +7,5 @@ api = Api(
     doc="/api/v1",
     version="1",
 )
+
+db = SQLAlchemy()